requirements.txt
Flask~=1.1.2
Flask-SQLAlchemy~=2.4.1
flask-login~=0.5.0
flask-wtf~=0.14.3
wtforms~=2.3.1
sqlalchemy~=1.3.16
gunicorn~=20.0.04
psycopg2~=2.8.5
python-telegram-bot~=12.7